gas价格调整

+++ ioncwalletsdk/src/main/java/org/ionc/wallet/constant/ConstanParams.java
@@ -1,11 +1,11 @@
 package org.ionc.wallet.constant;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public class ConstanParams {
     public static final String DB_NAME = "ionchainwallet";
     /**
      * gas 值,以太坊最小值
      */
-    public final static BigInteger GAS_MIN = BigInteger.valueOf(21000);
+    public final static BigDecimal GAS_MIN = BigDecimal.valueOf(21000);
 }
